requirements.txt
requests
bs4
selenium
msedge-selenium-tools
html5lib